Prevent race conditions involving SIGINT handling
Until now if the user pressed Control+C whilst the program was cleaning up (for example waiting for pending packets) the program would die inmediately. This patch renames  the globals control_c_pressed and control_c_pressed_again to much clearer tx_done and rx_done. It also adds propper handling of control_c count handling in the handler itself as a result the wait will now only be cancelled when pressing control+c a second time during the program life instead of on the first control+c if the user pressed it when we are already doing the scan.